allow custom config file source in the form of an array
The $source argument is not flexible enough as it is. It currently
forces the source to be unique.

Sometimes we'd like to be able to specify an array of possible sources.

So we'll use the $source directly with the file resource without
enforcing its format.

WATCH OUT: this could break currently deployed apache::config::file
resources, so one should change occurences of this resource's argument
from:
  source => "modules/site-apache/blah"
to a full source specification string:
  source => "puppet:///modules/site-apache/blah"