config: Implement config files as links
Some packages in Debian manage an Apache config file in their own /etc
directory and create a link to that file in Apache's conf.d directory.
One example package that does this is nagios3.

Introduce a new argument, $target, that can be used to specify the link
destination. When using $target, the value to $ensure must also be
'link' to make the file into a symbolic link. This is to make it
explicit when we want the apache::config::file to behave differently
than the normal behaviour of creating a plain text file in conf.d.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>