Avoid setting source or content when trying to remove apache::config::file
Currently, when trying to remove a config file that is not present in
one of the source directories, you get an error about puppet not being
able to find the file, which kills the puppet run.

We need a special case for when $ensure is set to 'absent' or 'purged'
so that we avoid setting the source argument.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>