make sure different routes go to different workers
Apache dispatches proxy requests to workers, which are named by the
url.
However, since we - especially in the context of proxying containers -
might actually have different socat processes providing a different
UNIX socket, we have to name the workers differently.

Otherwise requests for different routes, that should go to different
backends keep ending up on the first worker that was initialized with
that name and thus all subsequent requests will end up on the same
worker although they should go to adifferent backend.